requirements.txt
pyirsdk
pyyaml
requests
customtkinter
